dashboard controller 테스트중 history 테스트중

+++ src/main/java/com/park/monitoring/mapper/DiskMapper.java
@@ -9,9 +9,9 @@ import java.util.List;
 public interface DiskMapper {
 
     List<Disk> selectAllDisk();
-    List<Disk> selectAllDiskByServerId(Long serverId);
-    Disk selectDiskById(Long id);
+    List<Disk> selectAllDiskByServerId(Integer serverId);
+    Disk selectDiskById(Integer id);
     int insertDisk(Disk disk);
     int updateDisk(Disk disk);
-    int deleteDisk(Long id);
+    int deleteDisk(Integer id);
 }
